+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## 0.0.59
+
+- Default Tab Fixed
+
+## 0.0.58
+
+- View -> Show Symbol -> Show Tabs
+
+## 0.0.57
+
+- Indentation Implemented
+
+## 0.0.56
+
+- Search & Replace Menu -> Fixed double click on results
+
+## 0.0.55
+
+- Search & Replace Menu -> Fixed casing in search
+
 ## 0.0.54
 
 - Implemented: